Updates and fixes
